event/SocketEvent: pass events as unsigned, not short

+++ src/PipeCaptureBuffer.cxx
@@ -18,7 +18,7 @@ PipeCaptureBuffer::~PipeCaptureBuffer()
 }
 
 void
-PipeCaptureBuffer::OnSocket(short)
+PipeCaptureBuffer::OnSocket(unsigned)
 {
     auto w = buffer.Write();
     if (!w.IsEmpty()) {
